[chore #163911078] switch and rename v1 to v2
