+++ run.py
@@ -1,6 +1,6 @@
 from flask import Flask
 
-from api.v1 import create_app
+from api.v2 import create_app
 
 env = "development"
